Fix warnings
This CL fixes some warnings found by MSVC++.

+++ starboard/shared/starboard/player/filter/audio_renderer_impl_internal.cc
@@ -308,8 +308,8 @@ void AudioRendererImpl::UpdateSourceStatus(int* frames_in_buffer,
   *is_playing = !paused_.load() && !seeking_.load();
 
   if (*is_playing) {
-    *frames_in_buffer =
-        frames_sent_to_sink_.load() - frames_consumed_by_sink_.load();
+    *frames_in_buffer = static_cast<int>(frames_sent_to_sink_.load() -
+                                         frames_consumed_by_sink_.load());
     *offset_in_frames = frames_consumed_by_sink_.load() % kMaxCachedFrames;
   } else {
     *frames_in_buffer = *offset_in_frames = 0;
@@ -424,8 +424,8 @@ void AudioRendererImpl::ProcessAudioData() {
     // There are still audio data not appended so schedule a callback later.
     SbTimeMonotonic delay = 0;
     if (kMaxCachedFrames - frames_in_buffer < kMaxCachedFrames / 4) {
-      int frames_to_delay =
-          kMaxCachedFrames / 4 - (kMaxCachedFrames - frames_in_buffer);
+      int frames_to_delay = static_cast<int>(
+          kMaxCachedFrames / 4 - (kMaxCachedFrames - frames_in_buffer));
       delay = frames_to_delay * kSbTimeSecond / decoder_->GetSamplesPerSecond();
     }
     process_audio_data_scheduled_ = true;
@@ -436,8 +436,8 @@ void AudioRendererImpl::ProcessAudioData() {
 bool AudioRendererImpl::AppendAudioToFrameBuffer() {
   SB_DCHECK(BelongsToCurrentThread());
 
-  int frames_in_buffer =
-      frames_sent_to_sink_.load() - frames_consumed_by_sink_.load();
+  int frames_in_buffer = static_cast<int>(frames_sent_to_sink_.load() -
+                                          frames_consumed_by_sink_.load());
 
   if (kMaxCachedFrames - frames_in_buffer < kFrameAppendUnit) {
     return false;
@@ -447,7 +447,8 @@ bool AudioRendererImpl::AppendAudioToFrameBuffer() {
 
   // When |playback_rate_| is 0, try to fill the buffer with playback rate as 1.
   // Otherwise the preroll will never finish.
-  float playback_rate_to_fill = playback_rate_ == 0.f ? 1.f : playback_rate_;
+  float playback_rate_to_fill =
+      playback_rate_ == 0.0 ? 1.f : static_cast<float>(playback_rate_);
   scoped_refptr<DecodedAudio> decoded_audio =
       time_stretcher_.Read(kFrameAppendUnit, playback_rate_to_fill);
   SB_DCHECK(decoded_audio);
